Fixed bugs and added mechanics to regular pong
Solved bugs such as the instance in which the ball would get stuck in a paddle or the side wall and ricochet repeatedly against the same surface.  Changed flipVelocity() function and several others to add the ability to change the ball's angle according to where on the paddle you hit it. Added a score text box and changed the program to not close after one single score. Ball and paddles reset to center after each point. Ball starts at a random -30 to 30 degree angle in the same left/right direction after each point.