further attempts at debugging
I believe the use of dt is causing a slight time inaccuracy due to a natural inability to move an object by fractions of pixels at a time. For example, in a single call of the ball's update function, dt is (for example) 0.009, and the ball would have to move 2.2409999... pixels exactly to stay perfectly synced to the beat. With this system I can't think of any way to make the ball line up perfectly with the edge of the paddle as required